fix make run

+++ shop/entrypoint.sh
@@ -16,13 +16,16 @@ crond &
 
 TARGET_DIR="/var/www/html"
 ZIP_FILE="/var/www/shopware.zip"
+INSTALLED_FILE="/var/www/html/shop.installed"
 
 ls -la $TARGET_DIR
-if [ -f "$ZIP_FILE" ]; then
+
+if [ ! -f "$INSTALLED_FILE" ] && [ -f "$ZIP_FILE" ]; then
     echo "Unpacking Shopware..."
     unzip -q "$ZIP_FILE" -d "$TARGET_DIR"
-    find $TARGET_DIR -type f -name '.lock' -delete
+    find $TARGET_DIR -type f -name '.temp.docker' -delete
     rm "$ZIP_FILE"
+    touch "$INSTALLED_FILE"
     echo "Unpack done."
 fi
